moved password.ts to shared folder
