+++ app/webFrontend/src/app/auth/reset/reset.component.ts
@@ -7,7 +7,7 @@ import {MatSnackBar} from '@angular/material';
 import {ActivatedRoute, Router} from '@angular/router';
 import {isNullOrUndefined} from 'util';
 import {matchPasswords} from '../../shared/validators/validators';
-import {pwPattern} from '../password';
+import {pwPattern} from '../../../../../../shared/validators/password';
 
 @Component({
   templateUrl: './reset.component.html',
